first_project/requirements.txt
asgiref==3.6.0
certifi==2023.5.7
charset-normalizer==3.2.0
Django==4.1.6
django-admin-interface==0.24.2
django-colorfield==0.8.0
django-crispy-forms==1.14.0
django-pwa==1.1.0
gunicorn==21.0.1
idna==3.4
packaging==23.1
Pillow==9.4.0
psycopg==3.1.9
psycopg2-binary==2.9.6
python-dotenv==1.0.0
requests==2.31.0
sqlparse==0.4.3
typing-extensions==4.7.1
tzdata==2023.3
urllib3==2.0.3
whitenoise==6.5.0
django-ckeditor